Temperature.app: fix temperature data URL
Hello All,

I am new to Window Maker. I wanted to install the Temperature.app dockapp
and found out that the ftp address that retrieves temperature data is
invalid. Investigating further I found that noaa has changed the address. I
have created a patch and am attaching it to this email.

I did not create a pull request, It wanted me to create a fork off of the
dockapps repository.

Maybe someone might find the dockapp useful and might use it.

Regards
Riza