Make AddMouseRegion's index unsigned
This was previously already done for wmbiff/wmgeneral: the index
argument is used as an index into an array, and is always only tested to
be below the upper bound. Passing in a negative value would pass that
test, and would result in an out-of-bounds access. All dockapps appear
to get that right, but libdockapp could see this violated by a (future)
user.